requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
flask
opencv-python
numpy
requests